@@ -34,47 +34,52 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <java.version>1.8</java.version>
+
+    <maven.compiler.target>17</maven.compiler.target>
+    <maven.compiler.source>17</maven.compiler.source>
 
     <assertj-core.version>3.8.0</assertj-core.version>
     <awaitility.version>1.4.0</awaitility.version>
     <carrotsearch.version>0.7.1</carrotsearch.version>
-    <codahale-metrics.version>3.0.2</codahale-metrics.version>
+    <dropwizard-metrics.version>4.2.23</dropwizard-metrics.version>
     <commons-fileupload.version>1.3.3</commons-fileupload.version>
     <commons-io.version>2.15.1</commons-io.version>
+    <ehcache.version>3.10.8</ehcache.version>
+    <easymock.version>5.2.0</easymock.version>
     <commons-lang.version>2.6</commons-lang.version>
     <embed-mongo.version>1.50.0</embed-mongo.version>
-    <esapi.version>2.3.0.0</esapi.version>
+    <esapi.version>2.5.3.1</esapi.version>
     <freemarker.version>2.3.29</freemarker.version>
     <guava.version>19.0</guava.version>
     <geronimo-javamail_1.4_mail.version>1.8.3</geronimo-javamail_1.4_mail.version>
     <itextpdf.version>5.5.6</itextpdf.version>
     <javax.inject.version>1</javax.inject.version>
     <javax-rs.version>2.0</javax-rs.version>
-    <jersey.version>2.36</jersey.version>
-    <jetty.version>9.4.48.v20220622</jetty.version>
+    <jersey.version>3.1.3</jersey.version>
+    <jetty.version>11.0.19</jetty.version>
     <jgit.version>3.6.2.201501210735-r</jgit.version>
     <jhipsterloaded.version>0.7</jhipsterloaded.version>
     <joda-time.version>2.8.2</joda-time.version>
-    <json-path.version>2.2.0</json-path.version>
+    <json-path.version>2.8.0</json-path.version>
     <jsoup.version>1.17.2</jsoup.version>
     <jsr305.version>3.0.0</jsr305.version>
-    <logback.version>1.2.9</logback.version>
-    <logstash-logback.version>6.6</logstash-logback.version>
+    <logback.version>1.4.14</logback.version>
+    <logstash-logback.version>7.4</logstash-logback.version>
     <magma.version>4.0.1</magma.version>
     <metrics-spring.version>3.0.0</metrics-spring.version>
-    <mongodb-driver.version>4.7.2</mongodb-driver.version>
+    <mongodb-driver.version>4.9.1</mongodb-driver.version>
     <nosqlunit.version>0.7.9</nosqlunit.version>
-    <obiba-commons.version>3.0.7</obiba-commons.version>
+    <obiba-commons.version>4.0-SNAPSHOT</obiba-commons.version>
     <opal.version>4.5.2</opal.version>
     <opencsv.version>2.3</opencsv.version>
     <poi.version>5.2.5</poi.version>
-    <protobuf.version>2.5.0</protobuf.version>
+    <protobuf.version>3.24.3</protobuf.version>
     <protobuf-java-format.version>1.2.1-obiba</protobuf-java-format.version>
     <rql-parser.version>0.3.3-obiba</rql-parser.version>
     <shiro.version>1.13.0</shiro.version>
     <shiro-extras.version>1.1.0</shiro-extras.version>
     <lang-tag.version>1.4.4</lang-tag.version>
+    <servlet-api.version>6.0.0.RC1-jbossorg-1</servlet-api.version>
     <jakarta.servlet>4.0.4</jakarta.servlet>
 
     <!-- maven plugins -->
