@@ -29,6 +29,7 @@ import org.bouncycastle.openssl.PEMWriter;
 import org.obiba.mica.micaConfig.service.MicaConfigService;
 import org.obiba.security.KeyStoreManager;
 import org.obiba.security.KeyStoreRepository;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.stereotype.Service;
 
 import com.google.common.base.Strings;
@@ -39,7 +40,7 @@ import static java.util.Optional.ofNullable;
  *
  */
 @Service
-public class KeyStoreService {
+public class KeyStoreService implements InitializingBean {
 
   public static final String SYSTEM_KEY_STORE = "system";
 
@@ -55,8 +56,8 @@ public class KeyStoreService {
 
   private final KeyStoreRepository keyStoreRepository = new KeyStoreRepository();
 
-  @PostConstruct
-  public void init() {
+  @Override
+  public void afterPropertiesSet() throws Exception {
     if(keystoresRoot == null) {
       keystoresRoot = new File(PATH_KEYSTORE.replace("${MICA_HOME}", System.getProperty("MICA_HOME")));
     }
